requirements.txt
# Python dependencies for Anus AI
# Core dependencies
openai>=1.0.0
anthropic>=0.5.0
langchain>=0.0.267
pydantic>=2.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.0.0
tqdm>=4.66.0
google-genai>=1.5.0

# Web and browser automation
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.8.5

# Document processing
pypdf>=3.15.0
python-docx>=0.8.11
openpyxl>=3.1.2
pillow>=10.0.0

# Code execution
jupyter-client>=8.3.0
nbformat>=5.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0

# API and web interface
fastapi>=0.103.0
uvicorn>=0.23.0
streamlit>=1.26.0
